Bugfix of gpu cases
type: bugfix
target: gpu
reason:

------

+++ mindspore/core/ir/dtype/type.cc
@@ -217,7 +217,7 @@ size_t GetTypeByte(const TypePtr &type_ptr) {
   }
 }
 
-int GetTypeId(const TypeId &type_id) { return type_id; }
+int64_t GetTypeId(const TypeId &type_id) { return static_cast<int64_t>(type_id); }
 
 bool Type::operator==(const Value &other) const {
   if (!other.isa<Type>()) {
